feat: restructure core elements by moving Node, Beam, and Frame classes to the elements module
