+++ python/core/elements/node.py
